(Home) Find problem spmv

+++ Parallel/sparse.c
@@ -153,7 +153,7 @@ void spmv(sparse_CSR A, double * x, double len, double * result, const int myid,
     }
     
     int M = A.ncols;
-    double * x_gathered_elements = malloc(M*sizeof(int)); // maximum size
+    double * x_gathered_elements = malloc(M*sizeof(double)); // maximum size
 
     MPI_Win win;
     MPI_Win_create(x, len*sizeof(double), sizeof(double), MPI_INFO_NULL, comm, &win);
